Temp buttons mostly working.  A bit of tweaking left. Still need to add indicator led

+++ thermoc.h
@@ -18,8 +18,8 @@ public:
  
     // Decorate the Thermocouple lib
     //
-    float read_f();
-    float read_c();
+    int read_f();
+    int read_c();
 
 private:
     thermoc(const thermoc&); // disable copy ctor
